RAG/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aioredis==1.3.1
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asgiref==3.8.1
astroid==2.15.8
async-timeout==4.0.3
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
backcall==0.2.0
beautifulsoup4==4.13.4
billiard==4.2.1
bs4==0.0.2
build==1.2.2.post1
CacheControl==0.14.3
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
channels==3.0.5
channels-redis==3.4.1
charset-normalizer==3.4.2
cleo==2.1.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
crashtest==0.4.1
cryptography==45.0.4
daphne==3.0.2
dataclasses-json==0.6.7
decorator==5.2.1
dill==0.4.0
distlib==0.3.9
Django==4.2.23
django-extensions==3.2.3
dulwich==0.21.7
exceptiongroup==1.3.0
faiss-gpu
fastapi==0.115.12
fastjsonschema==2.21.1
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.5.1
greenlet==3.2.3
h11==0.16.0
hf-xet==1.1.3
hiredis==3.2.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.0
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.7.0
incremental==24.7.2
iniconfig==2.1.0
installer==0.7.0
ipython==7.34.0
isort==5.13.2
jaraco.classes==3.4.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
keyring==24.3.1
kombu==5.5.4
langchain==0.3.25
langchain-community==0.3.25
langchain-huggingface==0.3.0
langchain-mistralai==0.2.10
langchain-text-splitters==0.3.8
langchain-ollama==0.3.3
langsmith==0.3.45
lazy-object-proxy==1.11.0
lxml==5.4.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mccabe==0.7.0
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.4.4
mypy_extensions==1.1.0
networkx==3.2.1
numpy==1.26.4
orjson==3.10.18
packaging==24.2
parso==0.8.4
pathtools==0.1.2
pdf2image==1.17.0
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.2.1
pkginfo==1.12.1.2
platformdirs==4.3.8
pluggy==1.6.0
poetry==1.8.5
poetry-core==1.9.1
poetry-plugin-export==1.8.0
prompt_toolkit==3.0.51
propcache==0.3.2
ptyprocess==0.7.0
py==1.11.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
pylint==2.17.7
pyOpenSSL==25.1.0
PyPDF2==3.0.1
pyproject_hooks==1.2.0
pytesseract==0.3.13
pytest==6.2.5
pytest-django==4.5.2
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
rag==0.0.20
ragclip==0.0.4
RapidFuzz==3.13.0
redis==3.5.3
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==4.1.0
service-identity==24.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
sqlparse==0.5.3
starlette==0.46.2
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.3
torch==2.2.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.4
trove-classifiers==2025.5.9.12
Twisted==25.5.0
txaio==23.1.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.3
vine==5.1.0
virtualenv==20.31.2
watchdog==0.10.4
wcwidth==0.2.13
wrapt==1.17.2
xattr==1.1.4
yarl==1.20.1
zipp==3.23.0
zope.interface==7.2
zstandard==0.23.0
slugify